migrete 修正

+++ database/migrations/2022_08_06_163657_create_reviews_table.php
@@ -20,7 +20,7 @@ class CreateReviewsTable extends Migration
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
             
-            $table->unsignedBigInteger('category_id');
+            $table->unsignedInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories');
             
             $table->DateTime('visited_at');
